rename action-message-dispatcher to remote-action-message-dispatcher
