+++ src/views/insights/dependencies.ts
@@ -6,8 +6,8 @@ import * as ReactDOM from 'react-dom';
 import { BrowserAdapter } from '../../common/browser-adapters/browser-adapter';
 import { TelemetryEventSource } from '../../common/extension-telemetry-events';
 import { initializeFabricIcons } from '../../common/fabric-icons';
-import { RemoteActionMessageDispatcher } from '../../common/message-creators/action-message-dispatcher';
 import { ContentActionMessageCreator } from '../../common/message-creators/content-action-message-creator';
+import { RemoteActionMessageDispatcher } from '../../common/message-creators/remote-action-message-dispatcher';
 import { StoreActionMessageCreatorFactory } from '../../common/message-creators/store-action-message-creator-factory';
 import { StoreProxy } from '../../common/store-proxy';
 import { BaseClientStoresHub } from '../../common/stores/base-client-stores-hub';
